# docs/api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Calvinist Parrot Conversational API",
    "version": "1.0.0",
    "description": "Versioned contracts for Calvinist Parrot chat and Counsel of Three QA. Browser identity is always resolved server-side from the Appwrite session or server-managed guest cookie; v1 clients must not send `userId`."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Calvinist Parrot origin"
    }
  ],
...
  "paths": {
    "/api/v1/chats": {
      "post": {
        "tags": [
          "Chat"
        ],
        "operationId": "createChat",
        "summary": "Create a chat",
        "description": "Creates a chat and persists the initial message. When `initialAnswer` is supplied, the answer and optional classification fields are also persisted so a QA result can continue as a chat.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRequest"
              },
              "example": {
                "initialQuestion": "What does justification by faith mean?",
                "denomination": "reformed-baptist",
                "clientChatId": "cm5h7x9k20000v8b4r3m2n1q0",
                "requestId": "f6bb8ec0-6d64-4c30-9f15-3eeebc61ff48"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chat created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateChatResponse"
                },
                "example": {
                  "chatId": "cm5h7x9k20000v8b4r3m2n1q0",
                  "messageId": "8b90be0d-497f-48cb-8c67-85c2e28f2e83",
                  "requestId": "f6bb8ec0-6d64-4c30-9f15-3eeebc61ff48"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the contract.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The current browser actor does not own this chat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested chat or request was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The request failed before a stream could begin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/qa": {
      "post": {
        "tags": [
          "QA"
        ],
        "operationId": "askQa",
        "summary": "Stream a Counsel of Three answer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QaRequest"
              },
              "example": {
                "question": "What is the doctrine of the Trinity?",
                "denomination": "reformed-baptist"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The QA stream started.",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/QaStreamEvent"
                },
                "description": "Newline-delimited JSON. Each line is one complete JSON object matching QaStreamEvent. Once streaming begins, runtime failures are reported in-band as `error` events because the HTTP status has already been sent.",
                "example": {
                  "type": "reviewed_answer",
                  "content": "The Bible teaches that the one true God eternally exists as Father, Son, and Holy Spirit."
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the contract.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The current browser actor does not own this chat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested chat or request was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The request failed before a stream could begin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/parrot-chat": {
      "post": {
        "tags": [
          "Legacy"
        ],
        "operationId": "legacyParrotChatPost",
        "summary": "Legacy combined chat operation",
        "deprecated": true,
        "description": "Deprecated compatibility shim. Migrate to the operation-specific v1 chat routes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegacyParrotChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Legacy create/stop JSON response or message NDJSON stream, depending on the request shape.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CreateChatResponse"
                    },
                    {
                      "$ref": "#/components/schemas/StopChatResponse"
                    }
                  ]
                }
              },
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/ChatStreamEvent"
                },
                "description": "Newline-delimited JSON. Each line is one complete JSON object matching ChatStreamEvent. Once streaming begins, runtime failures are reported in-band as `error` events because the HTTP status has already been sent."
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the contract.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "403": {
            "description": "The current browser actor does not own this chat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "404": {
            "description": "The requested chat or request was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "409": {
            "description": "Legacy request ID conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "500": {
            "description": "The request failed before a stream could begin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Legacy"
        ],
        "operationId": "legacyParrotChatGet",
        "summary": "Legacy chat history",
        "deprecated": true,
        "parameters": [
          {
            "name": "chatId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 191
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Legacy chat history response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyGetChatResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the contract.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "403": {
            "description": "The current browser actor does not own this chat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "404": {
            "description": "The requested chat or request was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "500": {
            "description": "The request failed before a stream could begin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          }
        }
      }
    },
    "/api/parrot-qa": {
      "post": {
        "tags": [
          "Legacy"
        ],
        "operationId": "legacyParrotQa",
        "summary": "Legacy Counsel of Three stream",
        "deprecated": true,
        "description": "Deprecated compatibility shim. Migrate to `POST /api/v1/qa`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegacyQaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Legacy NDJSON payload served with its historical text media type.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaStreamEvent"
                },
                "description": "Newline-delimited JSON. Each line is one complete JSON object matching QaStreamEvent. Once streaming begins, runtime failures are reported in-band as `error` events because the HTTP status has already been sent."
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the contract.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "403": {
            "description": "The current browser actor does not own this chat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "404": {
            "description": "The requested chat or request was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          },
          "500": {
            "description": "The request failed before a stream could begin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Deprecation": {
                "description": "RFC 9745 date indicating when this compatibility endpoint was deprecated.",
                "schema": {
                  "type": "string",
                  "example": "@1785024000"
                }
              },
              "Link": {
                "description": "Migration documentation identified by the deprecation link relation.",
                "schema": {
                  "type": "string",
                  "example": "</api/v1/docs>; rel=\"deprecation\"; type=\"text/html\""
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "description": "A request-level error returned before streaming begins."
      },
      "CreateChatRequest": {
        "type": "object",
        "properties": {
          "initialQuestion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000,
            "description": "The initial question or message from the user."
          },
          "initialAnswer": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100000,
            "description": "An existing answer used to continue a QA result as a chat."
          },
          "denomination": {
            "type": "string",
            "enum": [
              "reformed-baptist",
              "presbyterian",
              "wesleyan",
              "lutheran",
              "anglican",
              "pentecostal",
              "non-denom"
            ],
            "description": "The theological tradition used to frame secondary doctrines."
          },
          "clientChatId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 191,
            "description": "An opaque resource identifier."
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 191,
            "description": "An idempotency identifier for one user request and its response."
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "subcategory": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "issueType": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "initialQuestion"
        ],
        "additionalProperties": false,
        "description": "Creates a chat and persists its initial user message."
      },
      "CreateChatResponse": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 191,
            "description": "An opaque resource identifier."
          },
          "messageId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 191,
            "description": "An opaque resource identifier."
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 191,
            "description": "An idempotency identifier for one user request and its response."
          }
        },
        "required": [
          "chatId",
          "messageId",
          "requestId"
        ],
        "additionalProperties": false,
        "description": "Identifiers needed to navigate to and continue the new chat."
      },
...
      "QaRequest": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000,
            "description": "A Bible or theology question for the Counsel of Three pipeline."
          },
          "denomination": {
            "type": "string",
            "enum": [
              "reformed-baptist",
              "presbyterian",
              "wesleyan",
              "lutheran",
              "anglican",
              "pentecostal",
              "non-denom"
            ],
            "description": "The theological tradition used to frame secondary doctrines."
          }
        },
        "required": [
          "question"
        ],
        "additionalProperties": false,
        "description": "Starts a streamed Counsel of Three response."
      },
      "QaStreamEvent": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "progress"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "type",
              "message"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "categorization"
              },
              "data": {
                "type": "object",
                "properties": {
                  "reformatted_question": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "subcategory": {
                    "type": "string"
                  },
                  "issue_type": {
                    "type": "string"
                  }
                },
                "required": [
                  "reformatted_question",
                  "category",
                  "subcategory",
                  "issue_type"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "type",
              "data"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "agent_responses"
              },
              "data": {
                "type": "object",
                "properties": {
                  "first_answer": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "second_answer": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "third_answer": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "first_answer",
                  "second_answer",
                  "third_answer"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "type",
              "data"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "calvin_review"
              },
              "content": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "type",
              "content"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "reviewed_answer"
              },
              "content": {
                "type": "string"
              }
            },
            "required": [
              "type",
              "content"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "refusal"
              },
              "content": {
                "type": "string"
              }
            },
            "required": [
              "type",
              "content"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "error"
              },
              "stage": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "type",
              "stage",
              "message"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "done"
              }
            },
            "required": [
              "type"
            ],
            "additionalProperties": false
          }
        ],
        "description": "One NDJSON line emitted by the Counsel of Three pipeline."
      },
...
    }
  }
}